+++ backend/src/main/scala/cromwell/backend/validation/PrimitiveRuntimeAttributesValidation.scala
@@ -65,7 +65,7 @@ class IntRuntimeAttributesValidation(override val key: String) extends
 
   override val womType = WomIntegerType
 
-  override protected def validateCoercedValue(womValue: WomInteger): ErrorOr[Int] = womValue.value.validNel
+  override protected def validateCoercedValue(womValue: WomInteger): ErrorOr[Int] = womValue.value.toInt.validNel
 
   override protected def typeString: String = "an Integer"
 }
@@ -75,7 +75,7 @@ class PositiveIntRuntimeAttributesValidation(override val key: String) extends
 
   override val womType = WomIntegerType
 
-  override protected def validateCoercedValue(womValue: WomInteger): ErrorOr[Int Refined Positive] = refineV[Positive](womValue.value).leftMap(NonEmptyList.one).toValidated
+  override protected def validateCoercedValue(womValue: WomInteger): ErrorOr[Int Refined Positive] = refineV[Positive](womValue.value.toInt).leftMap(NonEmptyList.one).toValidated
 
   override protected def typeString: String = "an Integer"
 }
